refactor(onboarding): change directory
