+++ apps/onboarding/src/app/measure/page.tsx
@@ -1,0 +1,3 @@
+export default function MeasurePage() {
+  return <div>Measure</div>;
+}
