+++ apps/onboarding/src/components/SVGWrapper.tsx
