+++ apps/onboarding/src/constants/alcohol.tsx
